Fix the casting integer to double within RecipeItem building

+++ app/src/main/java/Entities/ItemDisplays/Mass.java
@@ -8,7 +8,7 @@ public class Mass implements RecipeItemDisplay {
      * @return
      */
     @Override
-    public String display(float quantity, Ingredient ingredient) {
+    public String display(double quantity, Ingredient ingredient) {
         String stringQuantity;
 
         // checks if the quantity ends in .0 or .00 (whole number)
@@ -17,7 +17,7 @@ public class Mass implements RecipeItemDisplay {
             stringQuantity = Integer.toString((int) quantity);
         } else {
             // Convert to a string for returning
-            stringQuantity = Float.toString(quantity);
+            stringQuantity = Double.toString(quantity);
         }
         return stringQuantity + "g of " + ingredient.name();
     }
